+++ internal/handlemap/handle_map_test.go
@@ -9,7 +9,7 @@
 
    Licensed under the MIT License <http://opensource.org/licenses/MIT>.
 
-   Copyright © 2020-2024 Microsoft Corporation. All rights reserved.
+   Copyright © 2020-2025 Microsoft Corporation. All rights reserved.
    Author : <blobfusedev@microsoft.com>
 
    Permission is hereby granted, free of charge, to any person obtaining a copy
@@ -50,7 +50,7 @@ func (suite *HandleMapSuite) SetupTest() {
 	suite.assert = assert.New(suite.T())
 }
 
-// mount failure test where the mount directory does not exists
+// mount failure test where the mount directory does not exist
 func (suite *HandleMapSuite) TestNewHandle() {
 	h := NewHandle("abc")
 	suite.assert.NotNil(h)